+++ common/bitutil/bitutil.go
