+++ common/bitutil/bitutil_test.go
