+++ common/imageutil/asciiraster.go
