+++ common/imageutil/asciiraster_test.go
@@ -10,7 +10,7 @@ import (
 	"strings"
 	"testing"
 
-	"devt.de/common/fileutil"
+	"github.com/krotik/dudeldu/common/fileutil"
 )
 
 import (
