+++ common/lockutil/lockfile_test.go
@@ -16,7 +16,7 @@ import (
 	"testing"
 	"time"
 
-	"devt.de/common/fileutil"
+	"github.com/krotik/dudeldu/common/fileutil"
 )
 
 const lfdir = "lockfiletest"
